+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   [#51](https://github.com/green-code-initiative/creedengo-javascript/pull/51) Add ESLint v9 and flat config support
+-   [#88](https://github.com/green-code-initiative/creedengo-javascript/pull/88) Add rule `@creedengo/avoid-gettint-size-collection-in-loop` (GCI3)
 
 ### Changed