+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2026-02-03
+
 ### Added
 
 - [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Add SonarQube Community 25.12+ and SonarQube 2025.6+ support
@@ -15,8 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove ESLint v7/v8 support
 - [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove ESLint flat-specific configuration
-- [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove SonarQube Community < 25.12, SonarQube < 2025.6 support
-- [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove Node.js < 22 support
+- [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove SonarQube Community &lt; 25.12, SonarQube &lt; 2025.6 support
+- [#99](https://github.com/green-code-initiative/creedengo-javascript/pull/99) Remove Node.js &lt; 22 support
 
 ## [2.2.0] - 2026-01-15
 
@@ -123,7 +125,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setup basic coding style tools
 - Write complete contributing guide
 
-[unreleased]: https://github.com/green-code-initiative/creedengo-javascript/compare/2.2.0...HEAD
+[unreleased]: https://github.com/green-code-initiative/creedengo-javascript/compare/3.0.0...HEAD
+[3.0.0]: https://github.com/green-code-initiative/creedengo-javascript/compare/2.2.0...3.0.0
 [2.2.0]: https://github.com/green-code-initiative/creedengo-javascript/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/green-code-initiative/creedengo-javascript/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/green-code-initiative/creedengo-javascript/compare/1.5.0...2.0.0